requirements.txt
Twisted
PyYAML
exrex
python-iptables==0.11.0
netifaces==0.10.4